requirements.txt
asgiref==3.5.2
Django==4.1
djangorestframework==3.13.1
et-xmlfile==1.1.0
openpyxl==3.0.10
python-dotenv==0.20.0
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1
